Fixed bug with min i64
Due to using int64 to represent number we were unable
to handle min i64 as it overflows when we analyse literal
on its own without context of negation. We now use u64
and keep prefix in mind when performing validation

+++ src/evaluator/evaluator.go
@@ -285,9 +285,9 @@ func (e *Evaluator) eval(n ast.Node, ctx *Context) (result any) {
 		case *types.Char:
 			return e.evalCharCast(n.Value)
 		case *types.Any:
-			return n.Value
+			return int64(n.Value)
 		case *types.Generic:
-			return n.Value
+			return int64(n.Value)
 		default:
 			panic("this is a compiler error. please report")
 		}
